Move zero_diag into more general utils directory
