+++ alibi_detect/cd/pytorch/utils.py
@@ -17,10 +17,6 @@ def activate_train_mode_for_dropout_layers(model: Callable) -> Callable:
     return model
 
 
-def zero_diag(mat: torch.Tensor) -> torch.Tensor:
-    return mat - torch.diag(mat.diag())
-
-
 def quantile(sample: torch.Tensor, p: float, type: int = 7, sorted: bool = False) -> float:
     """ See https://wikipedia.org/wiki/Quantile#Estimating_quantiles_from_a_sample """
     N = len(sample)
