+++ alibi_detect/utils/pytorch/__init__.py
@@ -1,6 +1,7 @@
 from .distance import mmd2, mmd2_from_kernel_matrix, squared_pairwise_distance, permed_lsdds
 from .kernels import GaussianRBF
 from .prediction import predict_batch, predict_batch_transformer
+from .misc import zero_diag
 
 __all__ = [
     "mmd2",
@@ -9,5 +10,6 @@ __all__ = [
     "GaussianRBF",
     "permed_lsdds",
     "predict_batch",
-    "predict_batch_transformer"
+    "predict_batch_transformer",
+    "zero_diag"
 ]
