+++ alibi_detect/utils/tensorflow/__init__.py
@@ -2,6 +2,7 @@ from .distance import mmd2, mmd2_from_kernel_matrix
 from .distance import relative_euclidean_distance, squared_pairwise_distance, permed_lsdds
 from .kernels import GaussianRBF
 from .prediction import predict_batch, predict_batch_transformer
+from .misc import zero_diag
 
 __all__ = [
     "mmd2",
@@ -11,5 +12,6 @@ __all__ = [
     "GaussianRBF",
     "permed_lsdds",
     "predict_batch",
-    "predict_batch_transformer"
+    "predict_batch_transformer",
+    "zero_diag"
 ]
